Rename crypto::keygen to crypto::poly.
